refactor(test): move contour tests to _tests_ to fix package workflow failure
